Simplify demo app
Split demo in three different packages and activities, so there is a clear sample of automatic, manual and debug mode.

Related to YM-30926
