+++ sample/app/src/main/res/layout/activity_manual_face_capture.xml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FaceCaptureActivity">
+    tools:context=".manual.ManualFaceCaptureActivity">
 
     <com.yoti.mobile.android.capture.face.ui.FaceCapture
         android:id="@+id/face_capture_view"
